undo message change

+++ index.js
@@ -19,7 +19,8 @@ client.once('ready', () => {
 });
 
 client.on('messageCreate', async message => {
-	if (message.channelId === process.env.VERIFY_CHANNEL || message.channelId === process.env.CHECKPOINT_CHANNEL) {
+	if (!client.application?.owner) await client.application?.fetch();
+	if (message.author.id !== client.application?.owner.id && (message.channelId === process.env.VERIFY_CHANNEL || message.channelId === process.env.CHECKPOINT_CHANNEL)) {
 		message.delete();
 	}
 });
